+++ HelloCOUPIES/src/main/java/de/coupies/demoapp/fragment/AbstractFragment.java
@@ -31,7 +31,7 @@ import de.coupies.framework.session.CoupiesSession;
 public abstract class AbstractFragment extends Fragment {
 
 	private static String PARTNERS_USER_TOKEN = null;
-	private static final String API_KEY = ""; // TODO: please insert your COUPIES API Key here
+	private static final String API_KEY = "OZUajGUfSg"; // TODO: please insert your COUPIES API Key here
 	private static final String API_LEVEL = "4";
 	
 	private static final float LONGITUDE = 6.958237F;
